snippd_agent/requirements.txt
# Cloud Build / Agent Engine uses Linux + Python 3.11 (see ADK Dockerfile).
# Keep this file minimal so pip has fewer conflicts (Code 3 = build / pip failure).
#
# CRITICAL: Deploy from a folder whose basename is a valid Python identifier
# (letters, digits, underscore only). Use `snippd_agent`, NOT `snippd-agent`.

google-cloud-aiplatform[adk,agent_engines]>=1.112.0,<2.0.0
google-adk==1.30.0
supabase>=2.3.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0

# Optional: tighten if Cloud Build still fails — uncomment one line at a time:
# protobuf>=4.25.3,<6.0.0
# pydantic>=2.7.0,<3.0.0
